Update __init__.py
Problem: When a customized prompt is passed, the first item of the list of choices begins right after the prompt. Better would be on a new line.

Solution: If a customized prompt is passed to the function, there should be a new line character added to the prompt, before the list of choices is concatenated to it.